+++ vector/src/main/java/im/vector/riotx/features/media/ImageContentRenderer.kt
@@ -36,6 +36,7 @@ import im.vector.riotx.core.utils.isLocalFile
 import kotlinx.android.parcel.Parcelize
 import timber.log.Timber
 import javax.inject.Inject
+import kotlin.math.min
 
 class ImageContentRenderer @Inject constructor(private val activeSessionHolder: ActiveSessionHolder,
                                                private val dimensionConverter: DimensionConverter) {
@@ -56,7 +57,8 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
 
     enum class Mode {
         FULL_SIZE,
-        THUMBNAIL
+        THUMBNAIL,
+        STICKER
     }
 
     fun render(data: Data, mode: Mode, imageView: ImageView) {
@@ -112,7 +114,8 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
             // Clear image
             val contentUrlResolver = activeSessionHolder.getActiveSession().contentUrlResolver()
             val resolvedUrl = when (mode) {
-                Mode.FULL_SIZE -> contentUrlResolver.resolveFullSize(data.url)
+                Mode.FULL_SIZE,
+                Mode.STICKER   -> contentUrlResolver.resolveFullSize(data.url)
                 Mode.THUMBNAIL -> contentUrlResolver.resolveThumbnail(data.url, width, height, ContentUrlResolver.ThumbnailMethod.SCALE)
             }
             // Fallback to base url
@@ -149,18 +152,27 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
         val maxImageHeight = data.maxHeight
         val width = data.width ?: maxImageWidth
         val height = data.height ?: maxImageHeight
-        var finalHeight = -1
         var finalWidth = -1
+        var finalHeight = -1
 
         // if the image size is known
         // compute the expected height
         if (width > 0 && height > 0) {
-            if (mode == Mode.FULL_SIZE) {
-                finalHeight = height
-                finalWidth = width
-            } else {
-                finalHeight = Math.min(maxImageWidth * height / width, maxImageHeight)
-                finalWidth = finalHeight * width / height
+            when (mode) {
+                Mode.FULL_SIZE -> {
+                    finalHeight = height
+                    finalWidth = width
+                }
+                Mode.THUMBNAIL -> {
+                    finalHeight = min(maxImageWidth * height / width, maxImageHeight)
+                    finalWidth = finalHeight * width / height
+                }
+                Mode.STICKER   -> {
+                    // limit on width
+                    val maxWidthDp = min(dimensionConverter.dpToPx(120), maxImageWidth / 2)
+                    finalWidth = min(dimensionConverter.dpToPx(width), maxWidthDp)
+                    finalHeight = finalWidth * height / width
+                }
             }
         }
         // ensure that some values are properly initialized